+++ feature/google/build.gradle
@@ -1,0 +1,13 @@
+apply from: "$rootProject.projectDir/library.gradle"
+
+dependencies {
+    implementation fileTree(dir: "libs", include: ["*.jar"])
+
+    implementation project(path: ':common')
+    implementation project(path: ':public')
+    implementation project(path: ':strings')
+
+    implementation 'androidx.core:core-ktx:1.3.1'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
+    implementation "com.android.billingclient:billing:$billingVersion"
+}
