+++ feature/google/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=purchases-android
+POM_ARTIFACT_ID=purchases-store-google
+POM_PACKAGING=aar
